Extract feature contracts into separate files
Split *Feature.kt files into dedicated Action, State, and Effect files
for better code organization and maintainability.
