+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/memos/presentation/MemosAction.kt
@@ -66,48 +66,3 @@ sealed interface MemosAction {
     val token: String,
   ) : MemosAction
 }
-
-/**
- * State for the Memos feature.
- */
-data class MemosState(
-  val memos: List<Memo> = emptyList(),
-  val isLoading: Boolean = false,
-  val errorMessage: String? = null,
-  val credentialsMode: Boolean = false,
-  val baseUrl: String = "",
-  val token: String = "",
-  val isOfflineMode: Boolean = false,
-) {
-  val hasCredentials: Boolean get() = baseUrl.isNotBlank() && token.isNotBlank()
-  val needsCredentials: Boolean get() = !isOfflineMode && !hasCredentials
-}
-
-/**
- * Side effects for the Memos feature.
- */
-sealed interface MemosEffect {
-  data object LoadCachedMemos : MemosEffect
-
-  data object LoadRemoteMemos : MemosEffect
-
-  data class SaveCredentials(
-    val baseUrl: String,
-    val token: String,
-  ) : MemosEffect
-
-  data object LoadCredentials : MemosEffect
-
-  data class CreateMemo(
-    val content: String,
-  ) : MemosEffect
-
-  data class UpdateMemo(
-    val name: String,
-    val content: String,
-  ) : MemosEffect
-
-  data class DeleteMemo(
-    val name: String,
-  ) : MemosEffect
-}
