+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/mode/presentation/ModeSelectionState.kt
@@ -1,0 +1,14 @@
+package space.be1ski.vibits.shared.feature.mode.presentation
+
+data class ModeSelectionState(
+  val showCredentialsDialog: Boolean = false,
+  val baseUrl: String = "",
+  val token: String = "",
+  val isValidating: Boolean = false,
+  val error: ModeSelectionError? = null,
+)
+
+enum class ModeSelectionError {
+  FILL_ALL_FIELDS,
+  CONNECTION_FAILED,
+}
